refactor: move get_python_class_from_class_name to helpers
